Refactor common spider code into a base class.
Use longest common subsequence matching for bostonsing recordings.
