+++ shapenotecds/shapenotecds/spiders/singing.py
@@ -1,19 +1,12 @@
 # -*- coding: utf-8 -*-
 import scrapy
 import json
-import sqlite3
-import difflib
 import os
 import re
-import csv
 
-def open_db():
-    conn = sqlite3.connect("../minutes.db")
-    conn.text_factory = str
-    return conn
+from spider_base import SpiderBase
 
-
-class SingingSpider(scrapy.Spider):
+class SingingSpider(SpiderBase):
     name = "singing"
     allowed_domains = ["shapenotecds.com"]
     # start_urls = (
@@ -21,7 +14,7 @@ class SingingSpider(scrapy.Spider):
     # )
 
     def start_requests(self):
-        conn = open_db()
+        conn = self.open_db()
         curs = conn.cursor()
 
         # reqs = []
@@ -34,8 +27,6 @@ class SingingSpider(scrapy.Spider):
 
         curs.close()
         conn.close()
-
-        # return reqs
 
 
     def parse(self, response):
@@ -50,10 +41,7 @@ class SingingSpider(scrapy.Spider):
         if m:
             filename = m.group(1) + '.csv'
             if os.path.isfile(filename):
-                print "<FILE: %s>" % filename
-                with open(filename, 'rb') as f:
-                    song_data = list(csv.reader(f))
-                    self.parse_section(url, song_data)
+                self.parse_csv(filename, url)
                 return True
 
     def parse_old(self, response):
@@ -107,57 +95,3 @@ class SingingSpider(scrapy.Spider):
                 song_data.append((pagenum, url))
 
             self.parse_section(response.url, song_data)
-
-    def parse_section(self, audio_url, song_data):
-        conn = open_db()
-        curs = conn.cursor()
-
-        curs.execute('SELECT id FROM minutes WHERE audio_url=?', [audio_url])
-        row = curs.fetchone()
-        minutes_id = row[0]
-
-        # make lists of recording urls and song ids
-        songs = []
-        pages = []
-        urls = []
-        for pagenum, url in song_data:
-            altpage = ''
-            if pagenum[-1:] in ('t', 'b'):
-                altpage = pagenum[:-1]
-            else:
-                altpage = pagenum + 't'
-
-            curs.execute("SELECT id FROM songs WHERE PageNum IN (?, ?)", [pagenum, altpage])
-            row = curs.fetchone()
-            if row:
-                song_id = row[0]
-                songs.append(song_id)
-                urls.append(url)
-                pages.append(pagenum)
-            else:
-                self.logger.warning("no song id: " + pagenum)
-
-        # make list of songs and ids from the minutes
-        minutes_songs = []
-        minutes_ids = []
-        curs.execute("SELECT id, song_id FROM song_leader_joins WHERE minutes_id=?", [minutes_id])
-        for id, song_id in curs:
-            if not minutes_songs or minutes_songs[-1] != song_id:
-                minutes_songs.append(song_id)
-                minutes_ids.append([id])
-            else:
-                minutes_ids[-1].append(id)
-
-        # get the longest subsequence from recordings and minutes
-        s = difflib.SequenceMatcher(a=songs, b=minutes_songs)
-        last_a = 0
-        for a, b, n in s.get_matching_blocks():
-            for pagenum, url in zip(pages[last_a:a], urls[last_a:a]):
-                self.logger.warning("skip: %5s %s" % (pagenum, url))
-            last_a = a+n
-            for pagenum, url, join_ids in zip(pages[a:a+n], urls[a:a+n], minutes_ids[b:b+n]):
-                for id in join_ids:
-                    curs.execute("UPDATE song_leader_joins SET audio_url=? WHERE id=?", (url, id))
-                    self.logger.info("update: %5s %5s %s" % (id, pagenum, url))
-
-        conn.commit()
